refactor(desktop): split app.go into endpoints, reveal, transfer state and transfer
app.go 1097 -> 340. Four destinations, two of which already had their tests
written:

- endpoints.go: which server and web app this build talks to, and the settings
  that decide it. endpoints_test.go already existed and already tested exactly
  what moved in.
- reveal.go: showing a received file to the user, and the name and path checks
  that run before either platform command is handed anything. reveal_test.go,
  same story.
- transferstate.go: the one-transfer-at-a-time state machine, including the
  close guard.
- transfer.go: driving a send or a receive end to end.

The test layout has been asserting these two seams for longer than the sources
have honored them.

The risk in this split is the Wails binding surface. Wails reflects over
*App's package-scoped method set to generate the frontend bindings, and a
method that moved to another file in the same package is still in that set, so
the generated TypeScript must not change at all. It does not:
frontend/wailsjs/go/main/App.d.ts is byte-identical and still declares 23
functions.

One trap this hit and the compiler caught: the repo imports the standard
library's runtime as goruntime because the Wails runtime package already owns
the name. Splitting the file dropped the alias, and the two files that use
goruntime.GOOS would not compile until it was restored.

Test plan
- frontend/wailsjs/go/main/App.d.ts byte-identical; still 23 bindings.
- go doc -all -u on package main: identical to the base worktree.
- The exported method set on *App, which is what Wails actually reflects over,
  diffed on its own: identical.
- go build, go vet and go test clean in desktop/.
- No file is named with a GOOS or GOARCH suffix, so nothing gained an implicit
  build constraint.
